feat: contextual processing enhancements
